+++ product-accounting-backend/src/main/java/com/ichzerowan/accounting/controller/PurchaseController.java
@@ -39,13 +39,17 @@ public class PurchaseController {
     @PostMapping("")
     Purchase addPurchase(@RequestBody PurchaseDto purchaseDto){
         Purchase newPurchase = new Purchase(purchaseDto.getDate());
+        newPurchase.updateTotal();
         return repository.save(newPurchase);
     }
 
     @PutMapping("/{id}")
     Purchase updatePurchase(@RequestBody PurchaseDto purchaseDto, @PathVariable Long id){
         return repository.findById(id).map(purchase -> {
+            if(purchase.isCompleted())
+                throw new ModificationNotAllowedException(Purchase.class, purchase.getId());
             purchase.setDate(purchaseDto.getDate());
+            purchase.updateTotal();
             return repository.save(purchase);
         }).orElseThrow(() -> new ObjectNotFoundException(Purchase.class, id));
     }
@@ -54,6 +58,8 @@ public class PurchaseController {
     void deletePurchase(@PathVariable Long id){
         Purchase purchase =  repository.findById(id)
                 .orElseThrow(() -> new ObjectNotFoundException(Purchase.class, id));
+        if(purchase.isCompleted())
+            throw new ModificationNotAllowedException(Purchase.class, purchase.getId());
         repository.delete(purchase);
     }
 
@@ -61,10 +67,14 @@ public class PurchaseController {
     Purchase addPurchaseProduct(@RequestBody PurchaseProductDto productDto, @PathVariable Long purchaseId){
         Purchase purchase =  repository.findById(purchaseId)
                 .orElseThrow(() -> new ObjectNotFoundException(Purchase.class, purchaseId));
+        if(purchase.isCompleted())
+            throw new ModificationNotAllowedException(Purchase.class, purchase.getId());
         Product product = productRepository.findById(productDto.getProductId())
-                .orElseThrow(() -> new ObjectNotFoundException(Product.class, purchaseId));
+                .orElseThrow(() -> new ObjectNotFoundException(Product.class, productDto.getProductId()));
         PurchaseProduct purchaseProduct = new PurchaseProduct(purchase, product, productDto.getCount());
         purchaseProductRepository.save(purchaseProduct);
+        purchase.updateTotal();
+        repository.save(purchase);
         return purchase;
     }
 
@@ -73,16 +83,27 @@ public class PurchaseController {
         PurchaseProductId purchaseProductId = new PurchaseProductId(purchaseId, productId);
         PurchaseProduct purchaseProduct = purchaseProductRepository.findById(purchaseProductId)
                 .orElseThrow(() -> new ObjectNotFoundException(PurchaseProduct.class, purchaseId));
+        Purchase purchase = purchaseProduct.getPurchase();
+        if(purchase.isCompleted())
+            throw new ModificationNotAllowedException(Purchase.class, purchase.getId());
         purchaseProduct.setCount(productDto.getCount());
         purchaseProductRepository.save(purchaseProduct);
-        return purchaseProduct.getPurchase();
+        purchase.updateTotal();
+        repository.save(purchase);
+        return purchase;
     }
 
     @DeleteMapping("/{purchaseId}/products/{productId}")
-    void deletePurchaseProduct(@PathVariable Long purchaseId, @PathVariable Long productId){
+    Purchase deletePurchaseProduct(@PathVariable Long purchaseId, @PathVariable Long productId){
         PurchaseProductId purchaseProductId = new PurchaseProductId(purchaseId, productId);
         PurchaseProduct purchaseProduct = purchaseProductRepository.findById(purchaseProductId)
                 .orElseThrow(() -> new ObjectNotFoundException(PurchaseProduct.class, purchaseId));
+        Purchase purchase = purchaseProduct.getPurchase();
+        if(purchase.isCompleted())
+            throw new ModificationNotAllowedException(Purchase.class, purchase.getId());
         purchaseProductRepository.delete(purchaseProduct);
+        purchase.updateTotal();
+        repository.save(purchase);
+        return purchase;
     }
 }
